+++ lib/formatters/credit_card_number_input_formatter.dart
@@ -250,17 +250,14 @@ class CardSystemData {
 }
 
 class _CardSystemDatas {
-  /// рекурсивно ищет в номере карты код системы, начиная с конца
-  /// нужно для того, чтобы даже после setState и обнуления данных карты
-  /// снова правильно отформатировать ее номер
   static CardSystemData? getCardSystemDataByNumber(
     String cardNumber, {
-    int? subscringLength,
+    int? substringLength,
   }) {
     if (cardNumber.isEmpty) return null;
-    subscringLength = subscringLength ?? cardNumber.length;
+    substringLength = substringLength ?? cardNumber.length;
 
-    if (subscringLength < 1) return null;
+    if (substringLength < 1) return null;
     Map? rawData;
     List<Map> tempSystems = [];
     for (var data in _data) {