+++ lib/formatters/phone_input_formatter.dart
@@ -650,18 +650,28 @@ class PhoneCodes {
     return list;
   }
 
-  /// рекурсивно ищет в номере телефона код страны, начиная с конца
-  /// нужно для того, чтобы даже после setState и обнуления данных страны
-  /// снова правильно отформатировать телефон
+  /// Removes a country code from a phone
+  static String removeCountryCode(
+    String phoneWithCountryCode,
+  ) {
+    final countryData = getCountryDataByPhone(phoneWithCountryCode);
+    if (countryData != null) {
+      phoneWithCountryCode = phoneWithCountryCode.replaceAll('+', '').trim();
+      return phoneWithCountryCode.substring(countryData.phoneCode!.length);
+    }
+    return phoneWithCountryCode;
+  }
+
   static PhoneCountryData? getCountryDataByPhone(
     String phone, {
-    int? subscringLength,
+    int? substringLength,
   }) {
     if (phone.isEmpty) return null;
-    subscringLength = subscringLength ?? phone.length;
+    phone = phone.replaceAll('+', '');
+    substringLength = substringLength ?? phone.length;
 
-    if (subscringLength < 1) return null;
-    var phoneCode = phone.substring(0, subscringLength);
+    if (substringLength < 1) return null;
+    var phoneCode = phone.substring(0, substringLength);
 
     var rawData = _data.firstWhereOrNull(
       (data) =>
@@ -674,7 +684,7 @@ class PhoneCodes {
     if (rawData != null) {
       return PhoneCountryData.fromMap(rawData);
     }
-    return getCountryDataByPhone(phone, subscringLength: subscringLength - 1);
+    return getCountryDataByPhone(phone, substringLength: substringLength - 1);
   }
 
   static List<PhoneCountryData> getAllCountryDatasByPhoneCode(